Fix Game Freeze When Quitting to Main Menu and Restarting a Run
The issue was that the pause menu sets the time scale to 0, but the time scale doesn't get reset when starting the next run. This is why doing tricks fixed the problem, as they set the timescale to .5 and then reset it back to 1. Fixed by setting the timescale to 1 at the start of every run.